Overloaded isParticipant, added as_werewolf stat, added leader anti leave and leader anti self kick

+++ src/commands/lobby/LeaveCommand.ts
@@ -1,6 +1,7 @@
 import {Message, TextChannel} from "discord.js";
 import BaseCommand from "../../abstracts/BaseCommand";
 import ParticipationService from "../../services/ParticipationService";
+import ErrorUtil from "../../utils/ErrorUtil";
 
 class LeaveCommand extends BaseCommand {
 	constructor() {
@@ -19,6 +20,10 @@ class LeaveCommand extends BaseCommand {
 			const participationService = new ParticipationService();
 
 			const channel = message.channel as TextChannel;
+
+			if (await participationService.isLeader(message.author, channel.parent!)) {
+				return ErrorUtil.throwError(message, "You can't leave, use the transfer spell first.");
+			}
 
 			await participationService.removeUser(message.author, channel.parent!);
 		} catch (error) {
